[build] Separate protobuf java compilation into a seperate module
Currently we compile the protobufs in both the kudu-client
and the kudu-subprocess. Doing this also combines them onto
the same classpath as the module itself.

This patch breaks those classes into a separate `kudu-proto`
module so that we can avoid duplicate work and ignore
the `kudu-proto` module in API compatibility and test
coverage checks.

This module is not published, so there is no net change to the final
artifacts.

Change-Id: Ib83db74c4a09cdd597bee90f5124ed804adac42d
Reviewed-on: http://gerrit.cloudera.org:8080/16953
Tested-by: Kudu Jenkins
Reviewed-by: Alexey Serbin <aserbin@cloudera.com>