The callbacks need information about current leader to not have to reach back into kuduraft
Summary: When a No-OP recieved or Leader detected callback is called,
we need to set information about the current leader for any reporting
or redirection for clients. This can need the implementation of the
callbacks to make a reverse call into kuduraft to get the
ConsensusState. However the ground could have shifted and thereby the
callback can either fail or act on wrong information. There are several
options to fix this. One is to check that ground has not shifted. The
other option which we implement here, is to make the callback have
enough information so as to not have to reach back into kuduraft. This
is based on what MySQL Raft needs today, but can evolve over time.

The only case where the leader_details will still have partial info, is
when a non-config member has become a LEADER. That is a temporary rare
situation

Test Plan: Since this is a callback, the real test is at the application
MySQL Raft level. On kuduraft side we ensured that the compilation works which
ensures that the plumbing is correct.

Reviewers:

Subscribers:

Tasks:

Tags: