[build] Fix occasional Gradle build failures
Occasionally, on a clean build environment often as a part of the
CI pipline, the C++ build will fail with cryptic often Protobuf
related failures. It turns out this is a result of multiple
./gradlew calls which depend on the kudu-proto module.
This change fixes the issue by defining a single CMake target
to build the kudu-proto jar.

Change-Id: Iefa02e52a74f8a60c3029188995574ffdb0d3d5c
Reviewed-on: http://gerrit.cloudera.org:8080/16968
Tested-by: Kudu Jenkins
Reviewed-by: Alexey Serbin <aserbin@cloudera.com>
Reviewed-by: Andrew Wong <awong@cloudera.com>