Link up Election Context from TransferLeader to EDCB
Summary:

Previously we added more metdata for EDCB, but it was limited to a
single election. This change pipes ElectionContext received from
TransferLeader through kudu rpcs so that it eventually arrives at the
new leader's EDCB.

In MySQL Raft, when EDCB is called and it determines that we need
another elections, the ElectionContext can be passed back to preserve
context.

Test Plan:

Tested with fbcode.

Reviewers: anirbanr-fb, bhatvinay, yashtc

Subscribers:

Tasks:

Tags:
Signed-off-by: Yichen <yichenshen@fb.com>