[master] turn off client location assignment by default
This patch turns off location assignment by default to clients
connecting to a Kudu cluster.  The assigned locations are cached,
but the way how locations assignment performed is resource consuming,
see [1] for details.

There aren't many benefits in assigning locations to clients so far:
the only nice property of a client with assigned location vs a client
with no location assigned is that former case the client run from a
particular location would choose tablet servers in the same location
if performing scan in ReplicaSelection::CLOSEST_REPLICA mode.

This is a first step towards addressing KUDU-3212.

[1] https://issues.apache.org/jira/browse/KUDU-3212

Change-Id: I78474ced0a0129b3f2b1add55f6f908a136106d0
Reviewed-on: http://gerrit.cloudera.org:8080/17024
Tested-by: Kudu Jenkins
Reviewed-by: Hao Hao <hao.hao@cloudera.com>
Reviewed-by: Grant Henke <granthenke@apache.org>