updating voter_distribution inside the committed config
Summary: The voter distribution is used by Flexi Raft to
get proper quorums to use for commit and quorum intersection.

Till now we had no mechanism to update the voter distribution.
With this patch we provide a direct update from outside to
update the voter distribution inside the committed config.

When new regions are to be added to the ring, first the new
voter distribution should be sent and updated before the voting
members are added. FlexiRaft has the protections to make sure
that the minimum quorum size rules will be enforced during leader
election.

Test Plan: This was tested on the MySQL raft side with a unittest.

Reviewers: yichenshen

Subscribers:

Tasks:

Tags: