txn_commit-itest: deflake TestCommitTasksReloadOnLeadershipChange
The test shows up on the flaky test dashboard as failing around 20% of
the time. As it turns out, transferring leadership by quiescing multiple
replicas can lead to flakiness if we happen to pick a lagging replica as
the new leader.

Instead of targeting a specific tablet server as the host of the new
leaders, we'll now just quiesce the old leader tablet server and stop
quiescing the other tablet servers.

I ran the test in DEBUG mode 100 times. Before this patch, it failed 16
times; with it, it passed 100/100 times.

Change-Id: I2b27864e72888367eb0af7de59e044a9e018c31b
Reviewed-on: http://gerrit.cloudera.org:8080/17031
Tested-by: Kudu Jenkins
Reviewed-by: Hao Hao <hao.hao@cloudera.com>