rpc: reduce logging when server is shutting down
TxnCommitITest.TestRestartingWhileCommittingAndDeleting exercises the
case where the client RPCs are sent to the tserver as the tserver is
shutting down. This results in several log messages like:

I0205 05:01:20.190737  1310 messenger.cc:384] service kudu.tserver.TabletServerAdminService not registered on TabletServer
I0205 05:01:20.190891  1310 messenger.cc:384] service kudu.tserver.TabletServerAdminService not registered on TabletServer
I0205 05:01:20.194700  1310 messenger.cc:384] service kudu.tserver.TabletServerAdminService not registered on TabletServer

until the RPC timeout is reached. This message only seems interesting to
clients in cases outside the normal operation of a cluster, e.g. if the
endpoint actually doesn't exist, versus if we're just starting up or
shutting down. This patch adjusts the logging to reflect that.

Change-Id: I943a192bc12e3bae534b9129ac4ea8d24404adc6
Reviewed-on: http://gerrit.cloudera.org:8080/17032
Reviewed-by: Hao Hao <hao.hao@cloudera.com>
Tested-by: Andrew Wong <awong@cloudera.com>