[gutil] EmplaceOrDie() to return reference
I found it's useful for EmplaceOrDie() to return a reference
to the object that is constructed in-place in the container.

To implement this change, I needed to use std::enable_if()
dance because EmplaceOrDie() is already used in Kudu codebase
for not only dictionary, but for set containers as well:
that's so in src/kudu/ranger/ranger_client.cc at least.

Change-Id: Ife1e07b7454078d3392555866fd32b89fac28af7
Reviewed-on: http://gerrit.cloudera.org:8080/17034
Tested-by: Alexey Serbin <aserbin@cloudera.com>
Reviewed-by: Andrew Wong <awong@cloudera.com>