[util] remove AutoReleasePool and cleanup related code
This patch removes AutoReleasePool class and changes just two places
where it was used to store heap-allocated objects.  With this patch,
objects of appropriate type are wrapped into std::unique_ptr and
stored in std::deque container.  Along the way, I did other minor
code cleanup.

The motivation for this change is reducing number of memory allocations
and removing unneeded complexity in the code.

Change-Id: I8adb5d08ec37e716baf524cb6c5f52366553883c
Reviewed-on: http://gerrit.cloudera.org:8080/17036
Tested-by: Alexey Serbin <aserbin@cloudera.com>
Reviewed-by: Hao Hao <hao.hao@cloudera.com>