[kuduraft] [proxy] [perf] [4/n] blocking ReadOps in log-cache
Summary:
The proxy peer needs to read from log-cache/log-file to reconstitute the ops for
the destination peer. The proxy peer may not have the ops in its local log yet.
In such a case, the proxy peer used to do a busy wait (with sleeps) until the
ops are available (or until a configured timeout). This PR adds a new API in
LogCache to block until the proxied ops are available (or timeout after
specified duration).

1. LogCache::BlockingReadOps(...) is the new API. It is similar to
LogCache::ReadOps(...), but with a blocking semantic
2. LogCache::lock_ is changed from a spinlock to a mutex lock. This lock is
anyways contended and it is better to use a mutex. It also makes it simpler to
use the same lock to block using a condvar
3. RaftConsensus::HandleProxyRequest(...) is changed to make use of the new API
and the busy loop is removed.

Test Plan:
Existing proxy tests in fbcode

Reviewers:

Subscribers:

Tasks:

Tags: