[transactions] add TxnSystemClient::CheckOpenTxnSystemTable()
This patch adds a new ClientCheckOpenTxnSystemTable() method to
TxnSystemClient.  The new method is targeting the use case in a
follow-up patch where txn system client is used to automatically
register transaction participants while serving write requests
arriving to tablet servers.

NOTE: as of now, it would be necessary to re-open the transaction
      system table when a new partition added (i.e. when txn IDs
      grow) to make sense of using this approach with the newly
      introduced ClientCheckOpenTxnSystemTable() method.
      I'm planning to address that in a follow-up patch.

Change-Id: If2e9b3cfc1566ff8ddbe56be980fb7d360759c0f
Reviewed-on: http://gerrit.cloudera.org:8080/17044
Reviewed-by: Andrew Wong <awong@cloudera.com>
Tested-by: Kudu Jenkins