[thirdparty] use system curl for cmake
Since building Kudu from source requires curl to be installed,
let's rely on that system package when building cmake.  In addition
to reducing the build time of the third-party components, it also
helps avoiding build failures in environments where multiple
OpenSSL versions are installed.

I guess a proper fix would be updating cmake configure/build scripts
and makefiles, but I'm not sure it's worth it: removing an extra
component from the build gives the benefit of shorter build times.

Change-Id: I10a22e1e864b19491177c25a352e2378a6f6dea3
Reviewed-on: http://gerrit.cloudera.org:8080/17053
Tested-by: Kudu Jenkins
Reviewed-by: Grant Henke <granthenke@apache.org>