[txn_commit-itest] use hash partitioned table for tests
This patch switches the test table auto-created by TestWorkload
to be of the hash partitioned type.  This allows for easier reasoning
what tablets are to receive write operations when inserting a few
test rows with adjacent integer keys.  This change is essential
for a follow-up patch https://gerrit.cloudera.org/#/c/17037/

Change-Id: I8c0bfa803e115d5731e873aa3972bff29f4e93a7
Reviewed-on: http://gerrit.cloudera.org:8080/17068
Tested-by: Alexey Serbin <aserbin@cloudera.com>
Reviewed-by: Andrew Wong <awong@cloudera.com>