[test] fix flake in TestConcurrentRepeatedCommitCalls
Before this patch, the TestConcurrentRepeatedCommitCalls scenario
of the TxnCommitITest sometimes failed with an error like below:

  Bad status: Illegal state: Failed to write to server: 206a8320b5ea4e4c80360fd3de898ab5 (127.0.93.65:36189): transaction ID 0 is not open: state: COMMITTED user: "slave"

Before this patch, I saw such an error in about 1/8 of all DEBUG build
runs if specifying --stress_cpu_threads=16.  After this patch, I saw
none after running the scenario with the same flags hundreds of times.

Change-Id: I39118b2b33503a18b77e4e4106979f141b5e5486
Reviewed-on: http://gerrit.cloudera.org:8080/17069
Tested-by: Alexey Serbin <aserbin@cloudera.com>
Reviewed-by: Andrew Wong <awong@cloudera.com>