test: add more natural test for KUDU-2233
I have a patch in-flight that touches an area of the code around where
we expect the infamous KUDU-2233 crash. Before merging it, I'd like to
ensure the graceful handling of this corruption is unaffected,
especially in cases where data has previously been corrupted and we've
just upgraded to a newer version of Kudu.

This patch adds such a test case, where data is corrupted but does not
result in a crash, and the tserver is restarted with bits that handle
corruption gracefully. In doing so, this patch also adds an off switch
to all of the guardrails we installed for KUDU-2233.

Change-Id: Icac3ad0a1b6bb9b17d5b6a901dc5bba79c0840fa
Reviewed-on: http://gerrit.cloudera.org:8080/17114
Reviewed-by: Alexey Serbin <aserbin@cloudera.com>
Tested-by: Andrew Wong <awong@cloudera.com>