[consensus] DriverType: rename REPLICA to FOLLOWER
I found it a bit confusing to have DriverType::REPLICA along with
DriverType::LEADER in the consensus::DriverType enum.  In fact, both
stand for Raft replica roles, where the former is for a follower and
the latter is for a leader replica.  To resolve the confusion, this
patch renames DriverType::REPLICA into DriverType::FOLLOWER.

This patch does not contain any functional modifications.

Change-Id: Id30d5df01f2ea192795aeb23e525399c6869b5d6
Reviewed-on: http://gerrit.cloudera.org:8080/17113
Tested-by: Kudu Jenkins
Reviewed-by: Bankim Bhavsar <bankim@cloudera.com>