[client-test] added WriteWhileRestartingMultipleTabletServers
While working on tests for multi-row transactions, I added a similar
scenario and found it a bit flaky.  So, I wanted to see if the same
is true for the non-transactional write operations.  It turns out
the latter is pretty stable, so I need to dig in to find the root
cause of the former.  Anyways, I think this is a good scenario to
add into client-test.cc, extending already existing scenario
ClientTest.TestWriteWhileRestarting to multi-replica case and going
through multiple restarts.  Another important detail is that the newly
added test scenario verifies the number of persisted rows in the end.

I also did a small touch-up of the code related to the utility method
ClientTest::CountRowsFromClient().

Change-Id: I95d1456dea2e6e2bb7d8b0c5d05e95798098710d
Reviewed-on: http://gerrit.cloudera.org:8080/17120
Tested-by: Alexey Serbin <aserbin@cloudera.com>
Reviewed-by: Andrew Wong <awong@cloudera.com>