[kuduraft] change ReadReplicatesInRange() to send additional context
Summary:
The said API is changed to send additional information. For now,
additional information sent is the host and port of the peer for which
the request is being built. This enables upper layers to make use of
these information without having to call back into kuduraft for fetching
the same inforamtion again (which leads to deadlocks in certain conditions)

Test Plan:
Built local tp2 checkouts and ensured that proper information is passed

Reviewers: arahut

Subscribers:

Tasks:

Tags: