KUDU-3248: Match C++ replica selection behavior of Java client
The C++ client currently uses a new random value every time
a replica is selected. Alternatively, the Java client uses a static
value that ensures the selection for a single process or application
remains deterministic. This patch adjusts the C++ implementation
to match the Java client behavior.

This is expected to be a good balance of efficient use of cache
memory and distribution of load across the replicas given a
single process will always get the same choice resulting in
cache hits for follow up scans, but separate processes
will get a potentially different random selection.

The reviews on the Java patch here have some good context
and discussion: https://gerrit.cloudera.org/#/c/12158/

Change-Id: Iaa55e88b4a222fabfaa7fa521c24482cc6816b04
Reviewed-on: http://gerrit.cloudera.org:8080/17129
Tested-by: Kudu Jenkins
Reviewed-by: Attila Bukor <abukor@apache.org>