[tool] KUDU-3226 Validate List of Masters kudu ksck
This patch checks for duplicate kudu masters provided in all of the
kudu command line tools which call ParseMasterAddresses(). Currently
ksck and rebalance (as rebalance calls ksck internally) are the only
commands which throw the reported stack trace as ksck constructs a
map internally for the kudu masters and bombs once it sees duplicate
UUIDs. Other commands do not report any issue currently with duplicate
masters.

The patch only does a string comparison check to detect and report all
the duplicate master(/s). If a user provides the IP address and hostname
of the same master, it won't be able to catch the duplicates. That
seems more of a deliberate attempt rather than a common mistake a user
could make and hence not considering that case.

Also wrote a simple test case to test this functionality. Decided not
to spin up a test cluster for this as this is more of client side check
and a master is not actually contacted.

This change got conflicted with another test i.e.
AdminCliTest.TestAuthzResetCacheIncorrectMasterAddressList. Updated this
test as well to by creating a new external mini cluster with three
masters and running the command in the test with just one master

Change-Id: I2f3b2b7dcf2ac78cb95cf43242651e3ce8fddf6f
Reviewed-on: http://gerrit.cloudera.org:8080/17141
Reviewed-by: Alexey Serbin <aserbin@cloudera.com>
Tested-by: Kudu Jenkins
Reviewed-by: Andrew Wong <awong@cloudera.com>