[test] KUDU-3266 Fix flakiness in dynamic_multi_master test
Flakiness was reported in dynamic_multi_master test after
the introduction of test for recovering dead master,
commit 4b4a8c0f2f.

See KUDU-3266 for the analysis.

This change wraps the check for row count under ASSERT_EVENTUALLY
to ensure the resumed master and the remaining master are given
a chance to communicate Raft messages and become up to date.

Tests:
- Reproduced the issue with ASAN build with dist-test.
- Verified no failures over 100 iterations with the fix
on ASAN build.

Change-Id: Ifac1d95707064b6ac2624d3f52336d6c39afd3c8
Reviewed-on: http://gerrit.cloudera.org:8080/17211
Tested-by: Bankim Bhavsar <bankim@cloudera.com>
Reviewed-by: Andrew Wong <awong@cloudera.com>
Reviewed-by: Mahesh Reddy <mreddy@cloudera.com>
Reviewed-by: Alexey Serbin <aserbin@cloudera.com>